dvbaudiosink: turn off async state changes
When async state change is turned on and we are going
to paused state we are waiting for preroll and only after
preroll state change happens.

Problem is that preroll sometimes doesn't happen at all
and pipeline is stuck until FLUSH or EOS event is received.

To fix this we turn off async state changes, so we
will change state immediately without waiting for preroll

Only dvbaudiosink is not prerolling as described:
http://forums.openpli.org/topic/38783-gstreamer10-pauseunpause-improvements/?view=findpost&p=504374