dvbvideosink: update streamtype only when it's changed
When we have multiple caps events, we don't need to
send VIDEO_STOP, VIDEO_SET_STREAMTYPE and VIDEO_PLAY
ioctl for each of them, if the streamtype is not changed.

It improves perfomances of adaptive streaming playback,
since we don't stop video when caps are changed.

It also fixes picture-freeze issue on some receivers,
which was caused by sending VIDEO_STOP, VIDEO_SET_STREAMTYPE
and VIDEO_PLAY in quick succession.